Improved user friendliness
# Major update to improve script friendliness and ease of use.

## Change log:
- Added -h help option for arguments and information.
- Allow for the use of flags for use in scripts.
- Checks to see if script is running as root.
- Makes sure certbot is installed on the system prior to running.
- Prompts user for input if flags are not set.
- Stops script if domain, email, and server path are not set.
- Added logic to test for existing serverKey.pem before moving to prevents errors.
- Updated README to reflect new instructions.
- added gitignore file.

+++ GetSSL.sh
@@ -2,26 +2,160 @@
 
 # Created by: David Nahodyl, Blue Feather
 # Contact: contact@bluefeathergroup.com
-# Date: 5/7/2019
-# Version: 0.4
 
 # Need help? We can set this up to run on your server for you! Send an email to 
 # contact@bluefeathergroup.com or give a call at (770) 765-6258
 
-# Change the domain variable to the domain/subdomain for which you would like
-# an SSL Certificate
-DOMAIN="fms.mycompany.com"
+function usage() {
+    cat <<USAGE
 
-# Change the contact email address to your real email address so that Let's Encrypt
-# can contact you if there are any problems #>
-EMAIL="myemail@mycompoany.com"
+    WARNING! THIS SCRIPT WILL RESTART FILEMAKER SERVER!
 
-# Enter the path to your FileMaker Server directory, ending in a slash 
-SERVER_PATH="/Library/FileMaker Server/"
+    Creates an SSL Certificate from the Let's Encrypt Certificate Authority (CA) to
+    encypt data in motion for FileMaker. Certbot requires that port 80 be forwarded
+    to your server.
+    
+    Must be run as root.
 
-#
-# --- you shouldn't need to edit anything below this line
-#
+    Options:
+        -d | --domain               Set the domain.
+        -e | --email                Set the contact email.
+        -s | --server-path          Set the path to the FileMaker Server directory.
+
+        --no-confirm                Skip confirmation. Suggested for use in scripts.
+
+	    -h, --help, --usage:    	Print this help message.
+
+    Domain:
+    --domain
+        Change the domain variable to the domain/subdomain for which you would like
+        an SSL Certificate
+        "fms.mycompany.com"
+
+    Email:
+    --email
+        Change the contact email address to your real email address so that Let's Encrypt
+        can contact you if there are any problems
+        "myemail@mycompoany.com"
+
+    Server Path:
+    --server-path
+        Enter the path to your FileMaker Server directory, ending in a slash 
+        "/Library/FileMaker Server/" 
+
+
+USAGE
+    exit 1
+}
+
+# Set flags
+echo
+while [ "$1" != "" ]; do
+	case $1 in
+	-d | --domain)
+		shift
+		DOMAIN=$1
+		# echo "Using domain: $DOMAIN"
+		;;
+
+	-e | --email)
+		shift
+		EMAIL=$1
+		# echo "Email: $EMAIL"
+		;;
+
+	-s | --server-path)
+		shift
+		SERVER_PATH=$1
+		# echo "Server path: $SERVER_PATH"
+		;;
+
+    --no-confirm)
+        NOCONFIRM=true
+        ;;
+
+	-h | --help | --usage)
+		usage
+		exit 1
+		;;
+
+    	*)
+		printf "\033[1;31mError: Invalid option!\033[0m\n"
+		echo "Use --help for usage"
+		exit 1
+		;;
+	esac
+	shift
+done
+
+# Checks to see if script is running as root.
+if [ "$EUID" -ne 0 ]
+  then echo "Please run as root"
+  exit
+fi
+
+# Check to see if certbot is installed.
+if ! type certbot > /dev/null; 
+then
+    printf "\033[1;31mError: Certbot could not be found\033[0m\n"
+    echo "Install Certbot https://certbot.eff.org"
+    exit 1
+fi
+
+# Check for arguements.
+if [ "$DOMAIN" = "" ]; then
+	read -p "Set your domain: " DOMAIN
+        if [[ $DOMAIN == "" ]];
+            then printf "\033[1;31mError: Domain not specified. Must enter domain.\033[0m\n" && exit 1
+        fi
+	echo
+fi
+
+if [ "$EMAIL" = "" ]; then
+	read -p "Set your Email: " EMAIL
+        if [[ $EMAIL == "" ]];
+            then printf "\033[1;31mError: Email not specified. Must enter email.\033[0m\n" && exit 1
+        fi
+	echo
+fi
+
+if [ "$SERVER_PATH" = "" ]; then
+	read -p "Set your Server Path. Press 'enter' for default. ('/Library/FileMaker Server/'): " SERVER_PATH
+    SERVER_PATH=${SERVER_PATH:-"/Library/FileMaker Server/"}
+        if [[ $SERVER_PATH == "" ]];
+            then printf "\033[1;31mError: Server Path not specified. Must enter Server Path.\033[0m\n" && exit 1
+        fi
+	echo
+fi
+
+# Confirm arguements
+if  [[ $NOCONFIRM == "" ]]; 
+then
+    while true; do
+        echo "Domain:       $DOMAIN"
+        echo "Email:        $EMAIL"
+        echo "Server Path:  $SERVER_PATH"
+        echo
+        read -p "Is the above information correct? Y/n: " YN
+        case $YN in
+            [Yy]* )
+                echo "Continueing..."
+                break;;
+
+            [Nn]* )
+                echo "Stopping script..."
+                exit 1
+                break;;
+            * )
+                echo "Please answer yes or no.";;
+        esac
+    done
+else
+    echo "Skipping Confirmation..."
+fi
+
+# testing e-brake
+# exit
 
 WEB_ROOT="${SERVER_PATH}HTTPServer/htdocs"
 
@@ -35,7 +169,12 @@ cp "/etc/letsencrypt/live/${DOMAIN}/privkey.pem" "${SERVER_PATH}CStore/privkey.p
 chmod 640 "${SERVER_PATH}CStore/privkey.pem"
 
 # Move an old certificate, if there is one, to prevent an error
-mv "${SERVER_PATH}CStore/serverKey.pem" "${SERVER_PATH}CStore/serverKey-old.pem"
+FILE=${SERVER_PATH}CStore/serverKey.pem
+if test -f "$FILE"; then
+    echo "$FILE exists. Moving to serverKey-old.pem to prevent an error."
+    mv "${SERVER_PATH}CStore/serverKey.pem" "${SERVER_PATH}CStore/serverKey-old.pem"
+fi
+
 
 # Remove the old certificate
 fmsadmin certificate delete
@@ -51,3 +190,7 @@ sleep 15s
 
 # Start FileMaker Server again
 launchctl start com.filemaker.fms
+
+echo
+echo "FileMaker Server should now be set to use TLS/SSL"
+echo
